Split Storybook up per component
This is a precursor to some refactoring we need for sku 11, which
includes an upgrade to Storybook 6.
