--- src/stories/.npmignore
@@ -1,1 +1,0 @@
-**/*
