+++ src/storybook/markdown/assets/testImage.png
