+++ src/storybook/markdown/images-external.mdx
