+++ src/storybook/markdown/images-internal.mdx
@@ -6,6 +6,6 @@ Relative image without prefix `assets/testImage.png`:
 
 ![Alt text](assets/testImage.png '=400x Rest of title')
 
-Root-relative image `/src/stories/markdowns/assets/testImage.png`:
+Root-relative image `/src/storybook/markdown/assets/testImage.png`:
 
-![Alt text](/src/stories/markdowns/assets/testImage.png '=x600 Rest of title')
+![Alt text](/src/storybook/markdown/assets/testImage.png '=x600 Rest of title')
