+++ src/storybook/register.ts
@@ -1,0 +1,2 @@
+import 'braid-design-system/reset';
+import 'loki/configure-react';
